+++ rocketgram/executors/aiohttp.py
@@ -12,7 +12,7 @@ from aiohttp.web import Server, ServerRunner, BaseRequest, TCPSite, Response
 
 from .executor import Executor
 from ..api import Request, GetMe, SetWebhook, DeleteWebhook
-from ..api import Update, UpdateType
+from ..api import Update, UpdateType, InputFile
 from ..errors import RocketgramRequestError
 from ..version import version
 
@@ -71,13 +71,16 @@ class AioHttpExecutor(Executor):
         return len(request.files()) == 0
 
     async def add_bot(self, bot: 'Bot', *, allowed_updates: Optional[List[UpdateType]] = None,
-                      drop_pending_updates: bool = False, suffix: str = None, set_webhook: bool = True,
+                      drop_pending_updates: bool = False, certificate: Optional[InputFile] = None,
+                      ip_address: Optional[str] = None, suffix: str = None, set_webhook: bool = True,
                       max_connections: int = None):
         """
 
         :param bot:
         :param allowed_updates:
         :param drop_pending_updates:
+        :param certificate:
+        :param ip_address:
         :param suffix:
         :param set_webhook:
         :param max_connections:
@@ -102,8 +105,8 @@ class AioHttpExecutor(Executor):
         logger.info('Added bot @%s', bot.name)
 
         if set_webhook or drop_pending_updates:
-            swh = SetWebhook(full_url, allowed_updates=allowed_updates, drop_pending_updates=drop_pending_updates,
-                             max_connections=max_connections)
+            swh = SetWebhook(full_url, certificate=certificate, ip_address=ip_address, allowed_updates=allowed_updates,
+                             drop_pending_updates=drop_pending_updates, max_connections=max_connections)
             await bot.send(swh)
             logger.debug('Webhook setup done for bot @%s', bot.name)
 
@@ -230,6 +233,7 @@ class AioHttpExecutor(Executor):
     @classmethod
     def run(cls, bots: Union['Bot', List['Bot']], base_url: str, base_path: str, *, host: str = 'localhost',
             port: int = 8080, webhook_setup: bool = True, webhook_remove: bool = True,
+            certificate: Optional[InputFile] = None, ip_address: Optional[str] = None,
             allowed_updates: Optional[List[UpdateType]] = None, drop_pending_updates: bool = False,
             signals: tuple = (signal.SIGINT, signal.SIGTERM), shutdown_wait: int = 10):
         """
@@ -241,6 +245,8 @@ class AioHttpExecutor(Executor):
         :param port:
         :param webhook_setup:
         :param webhook_remove:
+        :param ip_address:
+        :param certificate:
         :param allowed_updates:
         :param drop_pending_updates:
         :param signals:
@@ -252,7 +258,8 @@ class AioHttpExecutor(Executor):
         executor = cls(base_url, base_path, host=host, port=port)
 
         def add(bot: 'Bot'):
-            return executor.add_bot(bot, allowed_updates=allowed_updates, drop_pending_updates=drop_pending_updates,
+            return executor.add_bot(bot, certificate=certificate, ip_address=ip_address,
+                                    allowed_updates=allowed_updates, drop_pending_updates=drop_pending_updates,
                                     set_webhook=webhook_setup)
 
         def remove(bot: 'Bot'):